pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-intensive-capstone-2025"
version = "0.1.0"
description = "Agentic AI based Adaptive Course Generator with a notebook UI and reusable Python package."
readme = "README.md"
requires-python = ">=3.10"

# ------------------------------------------------------------------
# Runtime dependencies 
# ------------------------------------------------------------------
dependencies = [
  "jupyterlab",
  "ipykernel",
  "pydantic>=2.3",
  "pydantic-settings>=2.0",   
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "google-adk",
  "google-adk[eval]",
  "rich",              # nice console logging for debugging / CLI
  "litellm",
  "pytest-asyncio"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "black",
  "mypy"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--strict-markers"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit (deselect with '-m \"not unit\"')"
]
[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B"]
ignore = []
exclude = [".venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = true
strict_optional = true
mypy_path = ["src"]
